Move secret_paths redaction to diagnostics dir
